requirements.txt
pyfast
numpy
h5py
pydicom
pillow
pylibjpeg
pylibjpeg-libjpeg
pylibjpeg-openjpeg
pyside2==5.15.2.1
